feat: Add ProblemPage component for displaying and filtering problems
feat: Implement LoginCard component for user authentication with validation

feat: Create SignUpCard component for user registration with password strength indicator

feat: Add seed-admin script to create an admin user if it doesn't exist
