+++ Frontend/src/api/api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-const BASE = import.meta.env.VITE_API_URL;
+const BASE = "http://localhost:8000/api/v1";
 const accessToken = localStorage.getItem("accessToken");
 
 export const login = (payload) => {
