+++ Frontend/src/views/contest/ContestCreatePage.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { createContest } from "../api/api.js";
+import { createContest } from "../../api/api.js";
 import { useNavigate } from "react-router-dom";
 
 export default function ContestCreatePage() {
