+++ Frontend/src/views/discussion/DiscussionPage.jsx
