+++ backend/src/app.js
@@ -8,13 +8,14 @@ const app = express();
 app.set('trust proxy', 1);
 
 const corsOptions = {
-  origin: process.env.CLIENT_URL || 'http://localhost:5173', 
+  origin: 'http://localhost:5173', 
   credentials: true, 
   allowedHeaders: ['Content-Type', 'Authorization'],
   methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
 };
 
 app.use(cors(corsOptions));
+// app.options('*', cors(corsOptions));
 
 
 app.use(express.json({limit: '50mb'}));
