+++ backend/src/config/db.config.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
 import { DATABASE_NAME } from "../constants.js";
+import { seedAdmin } from "../scripts/seed-admin.js";
 import dotenv from "dotenv";
 dotenv.config();
 
@@ -8,6 +9,14 @@ const connectDB = async () => {
         console.log("Connecting to MongoDB...");
         console.log(`MongoDB URI: ${process.env.MONGO_URI}${DATABASE_NAME}`);
         const connectionInstance = await mongoose.connect(`${process.env.MONGO_URI}${DATABASE_NAME}`);
+
+        // seedAdmin()
+        // .then(() => {
+        //     console.log("Admin is Seeded");
+        // })
+        // .catch(() => {
+        //     console.log("Error while seeding the admin");
+        // })
 
         console.log(`MongoDB connected: ${connectionInstance.connection.host}`);
