--- server/.dockerignore
@@ -1,5 +1,0 @@
-node_modules
-npm-debug.log
-Dockerfile
-.dockerignore
-.git
