--- server/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "singleQuote": true,
-  "printWidth": 80,
-  "tabWidth": 2,
-  "semi": true
-}
