--- server/Dockerfile
@@ -1,12 +1,0 @@
-FROM node:22
-
-WORKDIR /app
-
-COPY package*.json ./
-RUN npm install --only=production
-
-COPY . .
-
-EXPOSE 8000
-
-CMD ["npm", "start"]
