--- server/docs/architecture.md
@@ -1,14 +1,0 @@
-
-Problem Saving in S3
-
-s3://leetcode-clone/
- ├── problems/
- │    ├── problem-uuid/
- │    │    ├── description.md
- │    │    ├── sample_testcases.json
- │    │    ├── hidden_testcases.json
- │    │    ├── solution.cpp
- │    │    └── editorial.md
- ├── users/
- │    ├── user-uuid/
- │    │    └── submission-uuid.cpp
