--- server/drizzle.config.ts
@@ -1,16 +1,0 @@
-import * as dotenv from 'dotenv';
-
-dotenv.config();
-
-if (!process.env.DATABASE_URL) {
-  throw new Error('DATABASE_URL environment variable is not set');
-}
-
-export default {
-  schema: './src/drizzle/schema.ts',
-  out: './migrations',
-  dialect: 'postgresql',
-  dbCredentials: {
-    url: process.env.DATABASE_URL,
-  },
-};
