--- server/src/api/auth/auth-types.ts
@@ -1,14 +1,0 @@
-export interface UserType {
-    _id : string,
-    username : string,
-    email : string,
-    fullName : string,
-    role : string,
-    refreshToken : string
-}
-
-export interface RefreshTokenResponse {
-  accessToken: string;
-  refreshToken: string;
-}
-
