--- server/src/loaders/aws.ts
@@ -1,11 +1,0 @@
-import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
-import env from "../config";
-
-export const s3 = new S3Client({
-  region: env.AWS_REGION!,
-  credentials: {
-    accessKeyId: env.AWS_ACCESS_KEY,
-    secretAccessKey: env.AWS_SECRET_KEY,
-  },
-});
