--- server/src/utils/asyncHandler.ts
@@ -1,11 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-
-const catchAsync = (fn: any) => {
-  return (req: Request, res: Response, next: NextFunction) => {
-    fn(req, res, next).catch((err: any) => {
-      next(err);
-    });
-  };
-};
-
-export default catchAsync;
