--- server/src/utils/constants.ts
@@ -1,7 +1,0 @@
-import env from '../config';
-
-export const apiPrefix = '/api' + `/${env.API_VERSION}`;
-export const loggerConfig = {
-  level: env.LOG_LEVEL || 'info',
-};
-
